rename definitions file
